requirements.txt
matplotlib==3.5.1
numpy==1.21.5
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow==10.2.0
pycocotools==2.0.7
Requests==2.31.0
torch==2.2.1
torchvision==0.17.1
